+++ common/lib/src/realm_types.dart
@@ -79,7 +79,10 @@ class RealmStateError extends StateError implements RealmError {
   RealmStateError(String message) : super(message);
 }
 
+/// An error type that describes a session-level error condition.
+/// /// {@category Sync}
 class SessionError extends SyncError {
+  /// If true the received error is fatal.
   final bool isFatal;
 
   SessionError(