+++ lib/src/native/realm_core.dart
@@ -190,8 +190,7 @@ class _RealmCore {
         final syncConfigPtr = _realmLib.invokeGetPointer(() => _realmLib.realm_flx_sync_config_new(config.user.handle._pointer));
         try {
           _realmLib.realm_sync_config_set_session_stop_policy(syncConfigPtr, config.sessionStopPolicy.index);
-          _realmLib.realm_sync_config_set_error_handler(
-              syncConfigPtr, Pointer.fromFunction(_syncErrorHandlerCallback), config.toPersistentHandle(), nullptr);
+          _realmLib.realm_sync_config_set_error_handler(syncConfigPtr, Pointer.fromFunction(_syncErrorHandlerCallback), config.toPersistentHandle(), nullptr);
           _realmLib.realm_config_set_sync_config(configPtr, syncConfigPtr);
         } finally {
           _realmLib.realm_release(syncConfigPtr.cast());
@@ -400,12 +399,12 @@ class _RealmCore {
     if (syncConfig == null) {
       return;
     }
-    if (syncConfig.errorHandlerCallback != null) {
+    if (syncConfig.sessionErrorHandler != null) {
       final sessionError = error.toSessionError();
-      syncConfig.errorHandlerCallback!(sessionError);
+      syncConfig.sessionErrorHandler!(sessionError);
     }
   }
-  
+
   SchedulerHandle createScheduler(int isolateId, int sendPort) {
     final schedulerPtr = _realmLib.realm_dart_create_scheduler(isolateId, sendPort);
     return SchedulerHandle._(schedulerPtr);
@@ -1836,6 +1835,21 @@ extension on Pointer<Utf8> {
         _realmLib.realm_free(cast());
       }
     }
+  }
+}
+
+extension on realm_sync_error {
+  SessionError toSessionError() {
+    final messageText = detailed_message.cast<Utf8>().toRealmDartString()!;
+    final SyncErrorCategory errorCategory = SyncErrorCategory.values[error_code.category];
+    final isFatal = is_fatal == 0 ? false : true;
+
+    return SessionError(
+      messageText,
+      isFatal,
+      category: errorCategory,
+      code: error_code.value,
+    );
   }
 }